refactor(fs): split virtio fs device
